# backend/api-doc.yaml
openapi: 3.0.0
info:
  title: Kulhad Flask Backend API
  version: 1.0.0
  description: |
    OpenAPI 3.0 specification generated from the current Flask backend blueprints.

    Authentication-protected endpoints use JWT bearer authentication.
    The computer vision endpoints upload an image to the backend, which then calls an external object-detection service internally.
servers:
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /api/production/detect:
    post:
      tags:
        - production
      summary: Detect kulhads from an uploaded image
      description: |
        Uploads an image to the backend, which stores it temporarily and invokes an external object-detection service.

        The current Flask route validates image uploads and returns normalized detection results.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Detection completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionResponse'
              example:
                success: true
                kulhad_count: 2
                detections:
                  - class: kulhad
                    confidence: 0.87
                    x: 212
                    y: 180
                    width: 110
                    height: 140
                  - class: kulhad
                    confidence: 0.76
                    x: 140
                    y: 120
                    width: 90
                    height: 100
        '400':
          description: Invalid upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionErrorResponse'
              examples:
                missingFile:
                  value:
                    success: false
                    error: missing_file
                    message: Image file is required
                emptyFile:
                  value:
                    success: false
                    error: empty_file
                    message: Uploaded image is empty
                invalidType:
                  value:
                    success: false
                    error: invalid_file_type
                    message: "Unsupported image format. Allowed formats: .bmp, .jpeg, .jpg, .png, .webp"
        '500':
          description: Detection or configuration failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionErrorResponse'
              examples:
                inference:
                  value:
                    success: false
                    error: inference_failed
                configuration:
                  value:
                    success: false
                    error: configuration_error
                    message: ROBOFLOW_API_KEY is not configured
  /detect-kulhad:
    post:
      tags:
        - production
      summary: Detect kulhads from an uploaded image
      description: Alias of `/api/production/detect`.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Detection completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionResponse'
        '400':
          description: Invalid upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionErrorResponse'
        '500':
          description: Detection or configuration failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    DetectionPrediction:
      type: object
      properties:
        class:
          type: string
          example: kulhad
        confidence:
          type: number
          format: float
          example: 0.87
        x:
          type: integer
          example: 212
        y:
          type: integer
          example: 180
        width:
          type: integer
          example: 110
        height:
          type: integer
          example: 140
      required:
        - class
        - confidence
        - x
        - y
        - width
        - height
    DetectionResponse:
      type: object
      properties:
        success:
          type: boolean
        kulhad_count:
          type: integer
        detections:
          type: array
          items:
            $ref: '#/components/schemas/DetectionPrediction'
      required:
        - success
        - kulhad_count
        - detections
    DetectionErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
        message:
          type: string
      required:
        - success
        - error
...